server/requirements.txt
fastapi
uvicorn
langgraph
google-genai
supabase
python-dotenv
livekit
livekit-api
pydantic
httpx
edge-tts
